requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.9
decorator==4.1.1
dj-database-url==0.4.2
Django==1.11.3
djangorestframework==3.6.3
feedparser==5.2.1
future==0.16.0
gunicorn==19.7.1
idna==2.5
oauthlib==2.0.2
pathlib2==2.3.0
pbr==3.1.1
psutil==5.2.2
psycopg2==2.7.3
python-twitter==3.3
pytz==2017.2
requests==2.18.2
requests-oauthlib==0.8.0
scandir==1.5
six==1.10.0
stevedore==1.25.0
thefuck==3.18
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
whitenoise==3.3.0
